Develop

+++ .setup/options/develop/web/vscode.sh
@@ -4,13 +4,25 @@ echo "=========================== vs code ==========================="
 COMMAND_NAME="code"
 if ! command -v $COMMAND_NAME &>/dev/null; then
     echo "$COMMAND_NAME could not be found. Setting up $COMMAND_NAME."
-    sudo apt-get install -y wget gpg
-    wget -qO- https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor >packages.microsoft.gpg
-    sudo install -D -o root -g root -m 644 packages.microsoft.gpg /etc/apt/keyrings
-    sudo sh -c 'echo "deb [arch=amd64,arm64,armhf signed-by=/etc/apt/keyrings/packages.microsoft.gpg] https://packages.microsoft.com/repos/code stable main" > /etc/apt/sources.list.d/vscode.list'
-    rm -f packages.microsoft.gpg
-    sudo apt update
-    sudo apt install code # or code-insiders
+
+    # shellcheck disable=SC1091
+    if [ -f "/etc/os-release" ]; then
+        . /etc/os-release
+        OS=$NAME
+
+        if [ "$OS" == "Ubuntu" ]; then
+            sudo apt-get install -y wget gpg
+            wget -qO- https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor >packages.microsoft.gpg
+            sudo install -D -o root -g root -m 644 packages.microsoft.gpg /etc/apt/keyrings
+            sudo sh -c 'echo "deb [arch=amd64,arm64,armhf signed-by=/etc/apt/keyrings/packages.microsoft.gpg] https://packages.microsoft.com/repos/code stable main" > /etc/apt/sources.list.d/vscode.list'
+            rm -f packages.microsoft.gpg
+            sudo apt update
+            sudo apt install code # or code-insiders
+
+        elif [ "$OS" == "Zorin OS" ]; then
+            sudo snap install code --classic
+        fi
+    fi
 else
     echo "$COMMAND_NAME install ok installed"
 fi
